Modul2/Postawy_SQLAlchemy_ORM/Lekcja 1/requirements.txt
python-dotenv==1.0.1
psycopg2==2.9.9
asyncpg==0.29.0
aiosqlite==0.20.0